refactor: extract OrientationInfo in dedicated model
